Do not clear navigation child from cache while transitioning
This commit adds an additional conditional check before deleting the
navigation child from the cache list. This is to prevent the child getting
deleted, when 'getChildNavigation' is called by it's parent.

Happens when navigating back from a child navigator to the root,
and inside 'onNavigationStateChange' the 'screenProps' is changed.

Fixes #5301.