fix: don't pop routeKeyHistory when child handles back action (#75)
## Motivation

When pressing back in a StackNavigator nested inside the SwitchRouter, the screen would get popped, but the routeKeyHistory of the SwitchRouter gets popped even though the SwitchRouter doesn't handle the back action. This results in the next back action skipping the last opened tab.

## Fix

Before popping the `routeKeyHistory` stack, we check that the history actually changed. I also added a test for this